requirements.txt
Flask==2.3.3
openai==1.21.0
python-docx==1.1.0
